+++ libAfterStep/hints.h
@@ -233,9 +233,9 @@ typedef struct ASSupportedHints
 /* this calculates offset from the anchor position to the left/top corner of the frame : */
 #define APPLY_GRAVITY(grav,pos,size,bw1,bw2) \
 { switch( (grav) )					         \
-{case  0 : (pos) -= ((size)>>1)+(bw1)    ; break ; /* Center/Forget */ \
- case  1 : (pos) -= (bw1)+(bw2)+(size)   ; break ; /* South/East */    \
- case  2 : (pos) -= (bw1)                 ; break ; /* Static */        \
+{case  0 : (pos) -= ((int)(size)>>1)+(int)(bw1)    ; break ; /* Center/Forget */ \
+ case  1 : (pos) -= (int)(bw1)+(int)(bw2)+(int)(size)   ; break ; /* South/East */    \
+ case  2 : (pos) -= (int)(bw1)                 ; break ; /* Static */        \
  default :                                 break ; /* North/West */    \
 }}
